Add 'on destroy' bonuses and clean code
Introduced a new 'on_destroy_bonus.c' file to handle the application exit tasks cleanly such as releasing used memory and closing window. Meanwhile, restructured the code by reducing redundant lines and simplifying the function parameters to enhance readability and maintainability. Some changes are also reflected in the Makefile where the 'IS_BONUS' flag is changed from 1 to 0, indicating the changes made are for the base project and not the bonus part. This commit makes the code cleaner and easier to understand.
