+++ bonus/incs/render_bonus.h
@@ -27,10 +27,10 @@ int				ray_collide_vert(t_data *data, double vector[2], \
 int				fill_color(t_uint_img *dst, uint32_t floor, uint32_t ceiling);
 int				render_to_window(t_data *data);
 void			init_angles(t_data *data);
-int				cast_vert(t_data *data,double start[2], t_angle_data *angle, \
-						t_ray *ray, t_map map);
-int				cast_horiz(t_data *data,double start[2], t_angle_data *angle, \
-						t_ray *ray, t_map map);
+int				cast_vert(t_data *data, double start[2], t_angle_data *angle, \
+						t_ray *ray);
+int				cast_horiz(t_data *data, double start[2], t_angle_data *angle, \
+						t_ray *ray);
 t_angle_data	angles_sum(t_angle_data a, t_angle_data b);
 
 #endif
